requirements.txt
google-genai>=1.0.0
openpyxl>=3.0.0
pandas>=1.3.0
configparser
Flask>=2.0.0
werkzeug>=2.0.0 